pyproject.toml
[project]
name = "tutor-docente"
version = "1.0.0"
description = "Tutor Docente Verbum - Agente de formação docente usando Google ADK"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=0.5.0",
    "pydantic>=2.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "domain", "scripts"]

[tool.hatch.build.targets.sdist]
include = [
    "core/",
    "domain/",
    "scripts/",
    "agent.py",
    "agent_factories.py",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
